backend/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
--index-url https://download.pytorch.org/whl/cu121

# Diffusers for SDXL support
diffusers>=0.24.0
transformers>=4.40.0
accelerate>=0.24.0
peft>=0.8.0
safetensors>=0.4.0
pillow>=10.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
xformers>=0.0.22
compel>=2.0.0

# Real-ESRGAN dependencies for AI upscaling
realesrgan>=0.3.0
basicsr>=1.4.2
opencv-python>=4.8.0
facexlib>=0.3.0
gfpgan>=1.3.8

# Database dependencies
prisma>=0.11.0
psycopg2-binary>=2.9.7
asyncpg>=0.29.0
python-dotenv>=1.0.0

# R2/S3 Storage dependencies
boto3>=1.34.0
botocore>=1.34.0
Pillow>=10.0.0  # For image processing and thumbnails

# Queue system dependencies
redis>=5.0.0
celery>=5.3.0  # Alternative queue system
websockets>=12.0  # For real-time updates 
psutil>=5.9.0  # For worker process management